feat: use random character function as utills and remove it package
