+++ go.work
@@ -6,5 +6,4 @@ use (
 	./logger
 	./testsuite
 	./middleware/http-mdl
-	./random
 )
