+++ utils/utils_test.go
@@ -1,4 +1,4 @@
-package random
+package utils
 
 import (
 	"testing"
